+++ deepclone.stub.php
@@ -47,9 +47,9 @@ namespace {
      */
     const DEEPCLONE_HYDRATE_PRESERVE_REFS = UNKNOWN;
 
-    function deepclone_to_array(mixed $value, ?array $allowed_classes = null): array {}
+    function deepclone_to_array(mixed $value, ?array $allowed_classes = null, bool $allow_named_closures = false): array {}
 
-    function deepclone_from_array(array $data, ?array $allowed_classes = null): mixed {}
+    function deepclone_from_array(array $data, ?array $allowed_classes = null, bool $allow_named_closures = false): mixed {}
 
     function deepclone_hydrate(object|string $object_or_class, array $vars = [], int $flags = 0): object {}
 }